+++ TutoringSystem/TutoringSystem.Domain/Repositories/IAvailabilityRepository.cs
@@ -7,10 +7,10 @@ namespace TutoringSystem.Domain.Repositories
     public interface IAvailabilityRepository
     {
         Task<bool> AddAvailabilityAsync(Availability availability);
-        Task<Availability> GetAvailabilityByIdAsync(int availabilityId);
-        Task<Availability> GetTodaysAvailabilityByTutorIdAsync(int tutorId);
-        Task<ICollection<Availability>> GetAvailabilitiesByTutorIdAsync(int tutorId);
-        Task<ICollection<Availability>> GetFutureAvailabilitiesByTutorIdAsync(int tutorId);
+        Task<Availability> GetAvailabilityByIdAsync(long availabilityId);
+        Task<Availability> GetTodaysAvailabilityByTutorIdAsync(long tutorId);
+        Task<ICollection<Availability>> GetAvailabilitiesByTutorIdAsync(long tutorId);
+        Task<ICollection<Availability>> GetFutureAvailabilitiesByTutorIdAsync(long tutorId);
         Task<bool> UpdateAvailabilityAsync(Availability updatedAvailability);
         Task<bool> DeleteAvailabilityAsync(Availability availability);
     }
